docs/assets/msds-lec5-fastapi/requirements.txt
kafka-python
fastapi[all]
python-dotenv
confluent_kafka
Faker
pydantic
configparser